MarkX/requirements.txt
# Dmitry v1.2 Requirements

# Core
requests>=2.28.0
python-dotenv>=1.0.0
pillow>=9.0.0
sounddevice>=0.4.6
numpy>=1.24.0

# Speech
pyaudio>=0.2.13
SpeechRecognition>=3.10.0
edge-tts>=6.1.0

# Vector Database
chromadb>=0.4.22

# For testing
pytest>=7.0.0
